Swap to XCFrameworks
