+++ ios/Gemfile
@@ -1,3 +1,3 @@
 source 'https://rubygems.org'
 
-gem 'cocoapods', '= 1.6.0'
+gem 'cocoapods', '= 1.10.0'
