+++ ios/libs/ACPMediaTV.xcframework/tvos-arm64_x86_64-simulator/Headers/ACPMediaConstants.h
@@ -1,0 +1,108 @@
+//
+//  ACPMedia.h
+//  Adobe Digital Marketing Suite -- iOS Application Measurement Library
+//
+//  Copyright 1996-2020. Adobe, Inc. All Rights Reserved
+//
+//  Media Version: 2.3.0
+
+#import <Foundation/Foundation.h>
+
+/**
+ * @brief These constant strings define the stream type of the main content that is currently tracked.
+ */
+
+/**
+ * @brief Constant defining stream type for VOD streams
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaStreamTypeVod;
+
+/**
+ * @brief Constant defining stream type for Live streams
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaStreamTypeLive;
+
+/**
+ * @brief Constant defining stream type for Linear streams
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaStreamTypeLinear;
+
+/**
+ * @brief Constant defining stream type for Podcast streams
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaStreamTypePodcast;
+
+/**
+ * @brief Constant defining stream type for Audiobook streams
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaStreamTypeAudiobook;
+
+/**
+ * @brief Constant defining stream type for AOD streams
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaStreamTypeAod;
+
+
+/**
+ * @brief These constant strings define standard metadata keys for video content.
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyShow;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeySeason;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyEpisode;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyAssetId;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyGenre;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyFirstAirDate;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyFirstDigitalDate;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyRating;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyOriginator;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyNetwork;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyShowType;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyAdLoad;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyMvpd;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyAuthorized;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyDayPart;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyFeed;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPVideoMetadataKeyStreamFormat;
+
+/**
+ * @brief These constant strings define standard metadata keys for audio content.
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAudioMetadataKeyArtist;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAudioMetadataKeyAlbum;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAudioMetadataKeyLabel;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAudioMetadataKeyAuthor;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAudioMetadataKeyStation;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAudioMetadataKeyPublisher;
+
+/**
+ * @brief These constant strings define standard metadata keys for ads.
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAdMetadataKeyAdvertiser;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAdMetadataKeyCampaignId;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAdMetadataKeyCreativeId;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAdMetadataKeyPlacementId;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAdMetadataKeySiteId;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPAdMetadataKeyCreativeUrl;
+
+/**
+ * @brief These constant strings define additional event keys that can be attached to media object.
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaKeyMediaResumed;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaKeyPrerollTrackingWaitingTime;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaKeyGranularAdTracking;
+
+
+/**
+ * @brief These constant strings define keys that can be attached to config object.
+ */
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaKeyConfigChannel;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaKeyConfigDownloadedContent;
+
+/**
+ * @brief These constant strings define standard player states.
+*/
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaPlayerStateFullScreen;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaPlayerStatePictureInPicture;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaPlayerStateClosedCaption;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaPlayerStateInFocus;
+FOUNDATION_EXPORT NSString* _Nonnull const ACPMediaPlayerStateMute;
